Consistent HP and damage types (#8167)
Remove deprecated HUDs and chat message handling.
Remove unused m_damage variable (compat break).
HP: s32 for setter/calculations, u16 for getter.

+++ src/client/client.h
@@ -243,7 +243,7 @@ public:
 	void clearOutChatQueue();
 	void sendChangePassword(const std::string &oldpassword,
 		const std::string &newpassword);
-	void sendDamage(u8 damage);
+	void sendDamage(u16 damage);
 	void sendRespawn();
 	void sendReady();
 
@@ -342,7 +342,7 @@ public:
 	bool mediaReceived()
 	{ return !m_media_downloader; }
 
-	u8 getProtoVersion()
+	u16 getProtoVersion()
 	{ return m_proto_ver; }
 
 	bool connectedToServer();
@@ -504,7 +504,7 @@ private:
 	// and aren't accurate. We simply just don't know, because
 	// the server didn't send the version back then.
 	// If 0, server init hasn't been received yet.
-	u8 m_proto_ver = 0;
+	u16 m_proto_ver = 0;
 
 	u16 m_playeritem = 0;
 	bool m_inventory_updated = false;
