+++ src/client/clientenvironment.h
@@ -53,7 +53,7 @@ struct ClientEnvEvent
 		//struct{
 		//} none;
 		struct{
-			u8 amount;
+			u16 amount;
 			bool send_to_server;
 		} player_damage;
 	};
@@ -115,7 +115,7 @@ public:
 		Callbacks for activeobjects
 	*/
 
-	void damageLocalPlayer(u8 damage, bool handle_hp=true);
+	void damageLocalPlayer(u16 damage, bool handle_hp=true);
 
 	/*
 		Client likes to call these
